Ant and genetic working well

+++ GroupPlanner.Application/Algorithms/ScheduleEntryDto.cs
@@ -6,7 +6,7 @@ namespace GroupPlanner.Application.Algorithms
     {
         public DateTime Date { get; set; }
         public SubtaskDto? Subtask { get; set; }
-        public double Hours { get; set; }
+        public int Hours { get; set; }
 
         public string? SubtaskDescription => Subtask?.Description;
         public string? TaskEncodedName => Subtask?.TaskEncodedName;
