+++ GroupPlanner.Application/DailyAvailability/DailyAvailabilityDto.cs
@@ -12,7 +12,7 @@ namespace GroupPlanner.Application.DailyAvailability
 
         public DateTime Date { get; set; }
 
-        public double AvailableHours { get; set; }
+        public int AvailableHours { get; set; }
     }
 
 }
